Modify cgroup checker to comply with context (#137)
* Modify cgroup checker to comply with context

* Add test to verify checker complies with context

* Pass context to consistentRead

consistentRead now wraps the file reader with a context and will return
after context done in between reads. Couldn't seem to find a simple way
to go about interrupting a Read call. This implementation can't
interrupt in the middle of a Read call, but can still return early if
there are any attempts left, or in between Read operations.

* Return err with ConvertSystemError

* ReaderWithContext

- Added short `ReaderTestSuite` to test proper reader functionality and
to test context compliance.
- Added `utils_test` to hook up utils tests to "go test" runner.